Turns out that sensor was a LSM303DLH and not a LSM303DLM
I modified the order in which the magnetometer axis were read since Y
and Z weren't swapped on the actual sensor that I was using, a
LSM303DLH but I had originally swapped them based on the LSM303DLM
datasheet.

I also updated the name of the sensor class to make it clearer what
sensor I have hooked up on my breadboard.  The LSM303DLM documentation
was removed and replaced with the corresponding LSM303DLH ones.