Add moving average filter to sensor inputs
I updated the FilterPlay and compass samples to utilize a moving
average filter to smooth out some of the noise in the accelerometers
and magnetometers.  I found the magnetometers to be noisier than the
the accelerometers so I used a larger window for those sensors (16 vs
4).

In the compass sample, the smoothing is enabled by default.  You can
press the 's' key to toggle it on and off.

In the FilterPlay sample, the smoothing is diabled by default.  You can
press the 'f' key to toggle it on and off.