Update code to work with ADXL345 & HMC5883L
I first updated the firmware running on the mbed to interface with the
ADXL345 accelerometer and HMC5883L magnetometer instead of the
LSM303DLH combined accelerometer and magnetometer part.
NOTE: I plan to update this code in the near future to remove the
      duplicate code between the ADXL345 and HMC5883L classes.
I now run the I2C bus at 400kHz instead of the previous 100kHz.

I updated the Processing samples to work with the new sensors.  This
mostly comprised of:
* Updating the calibration data to deal with the range of values seen
  from the new sensors.  This deals with differences in resolution and
  zero bias error.
* Permuting the axis for determining orientation.  The new sensors
  don't have their axis arranged in the same directions as the previous
  ones...they don't even have them arranged the same with respect to
  the others sensors on the same Sparkfun 9DoF board.

I also removed the sensor orientation option from the compass sample as
I no longer needed it.  Previously I could switch the sensor from lying
flat and standing up on its edge.  This is no longer interesting.