Oversample ADX345 accelerometer
I modified the code to sample the accelerometer at 3200Hz, average
every 32 samples and then return the average along with the
magnetometer readings which are still only sampled at 100Hz.

I had hoped that this would reduce the noise I was seeing in the
accelerometer data but it didn't.  I think there is some low
frequency noise on the analog power supply which might be leading
to this.  I will get a 10uF capacitor and try adding that to see if
it helps.