Return 18-bit sums for accelerometer readings
Originally I was just reading the accelerometer sensor at 100Hz and
returning the 13-bit results.  Then I switched to sampling at 3200Hz
and returning the resulting 13-bit average.  I now send the complete
18-bit sum back to the PC.

I modified the calibration data for the sample to account for this
higher resolution and also extended the moving average window applied
to the accelerometers to match the magnetometer width (4 - 16).