Allow zeroing of yaw in orientation sample
The user can now press the space bar to zero out the yaw.  This allows
a demo to be started with the front of the device facing the user and
the UI of the sample will show the same orientation on the screen even
though north may not be directly to the left of the front face.