Add config file support to compass sample
The compass parameters like serial port name and sensor min/max values
are now stored in a .config file.  The base filename for the config
file is based on the USER environment variable.  For me the filename
ends up being adamgreen.config.  The adamgreen.config file in this
commit is an example of the supported settings.