Add rplidar sample
This is the initial Processing sample that I wrote for Xandon to use
for testing and integrating the RoboPeak LIDAR into his robot.  It's
still a bit rough around the edges but good enough to kickstart
Xandon's work.

The code under the rplidar/firmware directory was code that I wrote for
an mbed to simulate the protocol spoken by an actual RPLIDAR so that I
could test my code when I didn't have an actual RPLIDAR at my house to
test with.