Initial commit of calibration sample
I pulled the code together in this Processing sample to include all
of the routines that I have been using so far to measure the
calibration data for a particular sensor.

As the program runs it dumps stats (mean and variance) to stats.csv and
just the mean for the gyro measurements to gyro.csv.

It displays the averages results of the gyro and accelerometer
readings.  Pressing the A key will dump the current accelerometer
readings to the console as well.

It displays the min/max readings seen so far for the magnetometer.
Pressing M dumps these min/max values to the console.

It displays the accumulated gyro measurements.  Pressing the G key
dumps the current accumulated measurements, clears the sum, and takes
a snapshot of the current gyro bias readings to aid in correcting the
gyro readings as they are summed.