Fix rotation zeroing support
Previously the user could press the SPACE key to zero out the rotation
by pulling the camera around in virtual space to face the blue edge of
the sensor.  This would break the compass display and the text in the
upper right hand corner.  This commit fixes the compass and text
display by rotating around the Y axis to always face the camera.