Remove obtuse quaternion interpolation
When I calculate the quaternion to represent the accelerometer /
magnetometer rotation, it could end up using an inverted axis from
the gyro based quaternion.  This would cause the Kalman filter to
slowly correct by performing a ~360 degree rotation.  I now flip the
sign of the accel/mag quaternion if the angle between it and the
predicted quaternion are obtuse.