Better support other camera images
Camera resolutions up to 640x480 should now be supported.  The
Processing documentation indicates that the size() function should
be called first in the setup() routine with constant args.  There
are scenarios such as export which require these constant arguments.
By the time the camera had a frame ready for me to capture its
resolution, it was too late to call the size() function.