Add ability to correct for magnetic declination
FilteredCompass can now correct for magnetic declination which is
the deviation of magnetic north from true north.

Your .config file just needs to have a line like this:
  compass.declinationCorrection=16,7,26
This is the correction of 16 degrees, 7 minutes, 26 seconds currently
required for the Seattle area.  The adamgreen.config which is part of
this commit shows an example of setting it correctly.