Downsample camera image
I have added downSample and scaleBackUp constants to the blob sample
that allows for bilinear downsampling of the video feed.  I tested
with a 5x downsample and it does eliminate a lot of noise so that even
the ball tracking worked quite well.  This commit leaves the downsample
disabled but it can be re-enabled easily by configuring the downSample
constant at the beginning of blob.pde