Push downsample down into BlobDetector
The amount of bilinear down sample can be specified in the BlobDetector
constructor.  Previously I was down sampling in the main sample code
but doing it here makes it easier for it to be pulled over into
Robot_MC.  The previous approach made it easy to view the down
sampled image to see what filtering Processing used.  Now the down
sampled image only exists during blob detection and then it is
discarded.